Static Obstacles ENVIRONMENT Output
At the start of the trolley's motion, it travelled steadily and without delay until it reached a distance with a gap of hitting a static obstacle. Now the path is planned immediately based on the obstacles in the environment making sure the trolley does not hit anything and finally reaches the till.