Add files via upload
For this simulation analysis, we have shown that the environment has 4 static obstacles placed at points; (3, 2), (3, 4), (5, 6), (7, 7), and the line path has been changed according to the obstacles without hitting them but also following the line path after crossing the obstacle. The trolley has reached the goal (8, 9). The black path that our latest APF is designed for is collision free and global optimisation, ideal for the management of mobile robots.

At the start of the trolley's motion, it travelled steadily and without delay until it reached a distance with a gap of hitting a static obstacle. Now the path is planned immediately based on the obstacles in the environment making sure the trolley does not hit anything and finally reaches the till.