Static & Dynamic Environment - Initial OUTPUT
Both static and dynamic obstacles are in such a retail environment. Now the trolley learns that the customer has finished shopping and ready to pay. So the trolley travels to the till without any disruption in view of the obstacles. The static obstacles (blue rectangle blocks) here mentioned are the gondolas and the dynamic obstacles (red rectangles blocks) are the customers inside the supermarket. The first green circle is indicated as the starting point of the queue in the till and the second green circle is the payment point (till) of the supermarket.

All the dynamic obstacles are directed with an arrow for easy understanding. The path of the trolley is indicated in yellow.