Static & Dynamic Environment Final OUTPUT
Finally, the path planned trolley reaches the till for payment and successfully completes an APF based path-planning. Now in this supermarket environment, the trolley has overcame both the static as well as dynamic obstacles leading its way to the till. 

According to the simulation and experimental test, this approach has been shown to be effective for the avoidance of complex obstacles in uncertain and unstructured environments, improving the efficiency of robotic path planning considerably, reducing not only time requirements, but also reducing the distance. A significant amount of simulation shows that the enhanced APF approach escapes local minima and oscillating motions very quickly. Besides that, the simulation outcome confirms that a more global optimum, nearly ideal, collision-free and protection approach towards its target can still be calculated according to general APF with our suggested approach to path planning.